fix: attempt to fix redirection in deployed environment
